Move authentication to separate package and write initial interface
