+++ project/app/src/main/java/com/chalmers/tda367/localfeud/util/RestClient.java
@@ -20,6 +20,7 @@ public class RestClient {
     private RestResponseHandler responseHandler;
 
     public RestClient(IResponseAction action){
+
         client = new AsyncHttpClient();
         responseHandler = new RestResponseHandler(action);
     }
